deprecate!: dismiss target_dev attr on volumes
This refactor is part of an effort to make the configuration less
bloated and less error prone.

The target_dev attribute will be removed in a later release.

BREAKING CHANGE: any value for target_dev attribute will be dismissed
and replaced by a value calculated from the ordering of the volumes.